[유재우] feature/dashboard: 모듈화의 늪에 빠지다...
